Editing

+++ src/product/router.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter
 from fastapi.params import Depends, Query
-from sqlalchemy import select, update, delete, func
+from pygments.lexer import default
+from sqlalchemy import select, update
+from sqlalchemy.testing.suite.test_reflection import users
 
 from src.auth.router import cur_user
 from src.db import async_session
@@ -12,7 +14,7 @@ router = APIRouter(
     tags=["Products"]
 )
 
-@router.get("/get_product/", summary="Get a list of products")
+@router.get("/get_all_products/", summary="Get a list of products")
 async def get_product_list(user: User = Depends(cur_user)):
     async with async_session() as session:
         query = select(ProductTable).order_by(ProductTable.created_at)
@@ -24,8 +26,20 @@ async def get_product_list(user: User = Depends(cur_user)):
         else:
             return {"status": 200, "Products": ans_pr}
 
+@router.get("/get_your_products/", summary="Get a list of your products")
+async def get_local_product_list(user: User = Depends(cur_user)):
+    async with async_session() as session:
+        query = select(ProductTable).filter_by(user_id=user.id).order_by(ProductTable.price)
+        result = await session.execute(query)
+        ans = result.unique().scalars().all()
+        ans_pr = [ProductVU.model_validate(p) for p in ans]
+        if not ans_pr:
+            return{"message": "There is not a single product"}
+        else:
+            return {"status": 200, "Your products": ans_pr}
+
 @router.get("/get_status_product/{status}/", summary="Get a list of products with status(sold/unsold)")
-async def get_product_list(status: Status, user: User = Depends(cur_user)):
+async def get_product_list_with_status(status: Status, user: User = Depends(cur_user)):
     async with async_session() as session:
         query = select(ProductTable).filter_by(status=status)
         result = await session.execute(query)
@@ -50,9 +64,9 @@ async def create_product(pr: ProductCreate = Depends(), user: User = Depends(cur
             return {"status": 422, "message": f"Enter your ID, your ID does not match the one you entered({pr_s.user_id})."}
 
 @router.put("/edit_product_info/{product_id}/", summary="Edit product")
-async def edit_product(product_id: int ,name: str = Query(max_length=40), price: int = Query(ge=1)):
+async def edit_product(product_id: int ,name: str = Query(max_length=40), price: int = Query(ge=1), user: User = Depends(cur_user)):
     async with async_session() as session:
-        query = select(ProductTable).filter_by(id=product_id)
+        query = select(ProductTable).filter_by(id=product_id, user_id=user.id)
         res = await session.execute(query)
         pr_s = res.scalars().all()
         # noinspection PyBroadException
@@ -81,7 +95,7 @@ async def edit_product(product_id: int ,name: str = Query(max_length=40), price:
 @router.put("/sale_product/{product_id}/", summary="Product Sale")
 async def product_sale(product_id: int, user: User = Depends(cur_user)):
     async with async_session() as session:
-        query = select(ProductTable).filter_by(id=product_id)
+        query = select(ProductTable).filter_by(id=product_id, user_id=user.id)
         res = await session.execute(query)
         pr_s = res.scalars().all()
         # noinspection PyBroadException
@@ -109,19 +123,23 @@ async def product_sale(product_id: int, user: User = Depends(cur_user)):
             }
 
 @router.delete("/delete_product/{product_id}/", summary="Delete one Product")
-async def delete_product(product_id: int, user: User = Depends(cur_user)):
+async def delete_product(product_id: int, user_id: int, user: User = Depends(cur_user)):
     async with async_session() as session:
         # noinspection PyBroadException
         try:
-            del_prod = await session.get(ProductTable, product_id)
-            await session.delete(del_prod)
-            await session.commit()
-            return {
-                "status": 200,
-                "message": f"Product with {product_id} deleted",
-                "Deleted it": f"{user.username} {user.surname}",
-                    }
+            if user_id == user.id:
+                del_prod = await session.get(ProductTable, product_id)
+                await session.delete(del_prod)
+                await session.commit()
+                return {
+                    "status": 200,
+                    "message": f"Product with {product_id} deleted",
+                    "Deleted it": f"{user.username} {user.surname}",
+                        }
+            else:
+                return{"status": 404, "message": "Please check column <user_id>"}
         except:
+            print(del_prod)
             return {
                 "status": 404,
                 "message": f"Product with {product_id} was not found"
